Added React Router and mock pages
