+++ frontend/src/App.tsx
@@ -1,163 +1,25 @@
-import axios from 'axios'
-import { useEffect, useState } from 'react'
+import { BrowserRouter as Router, Switch, Route } from 'react-router-dom'
 
-import AppBar from '@material-ui/core/AppBar'
-import Button from '@material-ui/core/Button'
-import TrackChanges from '@material-ui/icons/TrackChangesOutlined'
-import Card from '@material-ui/core/Card'
-import CardActions from '@material-ui/core/CardActions'
-import CardContent from '@material-ui/core/CardContent'
-import CardMedia from '@material-ui/core/CardMedia'
-import CssBaseline from '@material-ui/core/CssBaseline'
-import Grid from '@material-ui/core/Grid'
-import Toolbar from '@material-ui/core/Toolbar'
-import Typography from '@material-ui/core/Typography'
-import { makeStyles } from '@material-ui/core/styles'
-import Container from '@material-ui/core/Container'
+import Layout from './Components/Layout'
+import SignUp from './Components/Auth/SignUp'
+import LogIn from './Components/Auth/LogIn'
 
 function App() {
-  const classes = useStyles()
-
-  const [data, setData] = useState<string | null>(null)
-
-  useEffect(() => {
-    async function fetchData() {
-      const res = await axios.get('/api/message')
-
-      setData(res.data)
-    }
-    fetchData()
-  }, [])
-
   return (
-    <>
-      <CssBaseline />
-      <AppBar position='relative' color='transparent'>
-        <Toolbar>
-          <TrackChanges className={classes.icon} />
-          <Typography variant='h6' color='inherit' noWrap>
-            Just The Next
-          </Typography>
-        </Toolbar>
-      </AppBar>
-
-      <main>
-        {/* Hero unit */}
-        <div className={classes.heroContent}>
-          <Container maxWidth='sm'>
-            <Typography
-              component='h1'
-              variant='h2'
-              align='center'
-              color='textPrimary'
-              gutterBottom
-            >
-              Just The Next
-            </Typography>
-            <Typography
-              variant='h5'
-              align='center'
-              color='textSecondary'
-              paragraph
-            >
-              Don't think about choosing what, go with just the next. {data}
-            </Typography>
-            <div className={classes.heroButtons}>
-              <Grid container spacing={2} justify='center'>
-                <Grid item>
-                  <Button variant='contained' color='primary'>
-                    Sign Up
-                  </Button>
-                </Grid>
-                <Grid item>
-                  <Button variant='outlined' color='primary'>
-                    Log In
-                  </Button>
-                </Grid>
-              </Grid>
-            </div>
-          </Container>
-        </div>
-        <Container className={classes.cardGrid} maxWidth='md'>
-          {/* End hero unit */}
-          <Grid container spacing={4}>
-            {[1].map((card) => (
-              <Grid item key={card} xs={12} sm={6} md={4}>
-                <Card className={classes.card}>
-                  <CardMedia
-                    className={classes.cardMedia}
-                    image='https://source.unsplash.com/random'
-                    title='Image title'
-                  />
-                  <CardContent className={classes.cardContent}>
-                    <Typography gutterBottom variant='h5' component='h2'>
-                      Feature 1
-                    </Typography>
-                    <Typography>This is the first Feature</Typography>
-                  </CardContent>
-                  <CardActions>
-                    <Button size='small' color='primary'>
-                      View
-                    </Button>
-                    <Button size='small' color='primary'>
-                      Edit
-                    </Button>
-                  </CardActions>
-                </Card>
-              </Grid>
-            ))}
-          </Grid>
-        </Container>
-      </main>
-      {/* Footer */}
-      <footer className={classes.footer}>
-        <Typography variant='h6' align='center' gutterBottom>
-          Just the Next
-        </Typography>
-        <Typography
-          variant='subtitle1'
-          align='center'
-          color='textSecondary'
-          component='p'
-        >
-          by Cruzeiro Dev Studio
-        </Typography>
-      </footer>
-      {/* End footer */}
-    </>
+    <Router>
+      <Switch>
+        <Route path='/signup'>
+          <SignUp />
+        </Route>
+        <Route path='/login'>
+          <LogIn />
+        </Route>
+        <Route path='/'>
+          <Layout />
+        </Route>
+      </Switch>
+    </Router>
   )
 }
-
-const useStyles = makeStyles((theme) => ({
-  icon: {
-    marginRight: theme.spacing(2),
-  },
-  heroContent: {
-    backgroundColor: theme.palette.background.paper,
-    padding: theme.spacing(8, 0, 6),
-  },
-  heroButtons: {
-    marginTop: theme.spacing(4),
-  },
-  cardGrid: {
-    paddingTop: theme.spacing(8),
-    paddingBottom: theme.spacing(8),
-  },
-  card: {
-    height: '100%',
-    display: 'flex',
-    flexDirection: 'column',
-  },
-  cardMedia: {
-    paddingTop: '56.25%', // 16:9
-  },
-  cardContent: {
-    flexGrow: 1,
-  },
-  footer: {
-    backgroundColor: theme.palette.background.paper,
-    padding: theme.spacing(6),
-  },
-}))
 
 export default App
