+++ frontend/src/Components/Auth/LogIn.tsx
@@ -1,0 +1,3 @@
+export default function LogIn() {
+  return <h1>Hello from LogIn</h1>
+}
