+++ frontend/src/Components/Auth/SignUp.tsx
@@ -1,0 +1,3 @@
+export default function SignUp() {
+  return <h1>Hello from SignUp</h1>
+}
